+++ packages/react-widgets/src/widgets/FormulaWidget.js
@@ -1,12 +1,10 @@
-import React, { useState, useEffect } from 'react';
-import { useSelector } from 'react-redux';
+import React from 'react';
 import { PropTypes } from 'prop-types';
 import { WrapperWidgetUI, FormulaWidgetUI } from '@carto/react-ui';
 import { getFormula } from '../models';
 import { AggregationTypes } from '@carto/react-core';
-import useSourceFilters from '../hooks/useSourceFilters';
-import { selectAreFeaturesReadyForSource } from '@carto/react-redux';
 import { columnAggregationOn } from './utils/propTypesFns';
+import useWidgetFetch from '../hooks/useWidgetFetch';
 
 /**
  * Renders a <FormulaWidget /> component
@@ -19,6 +17,7 @@ import { columnAggregationOn } from './utils/propTypesFns';
  * @param  {AggregationTypes} props.operation - Operation to apply to the operationColumn. Must be one of those defined in `AggregationTypes` object.
  * @param  {Function} [props.formatter] - Function to format each value returned.
  * @param  {boolean} [props.animation] - Enable/disable widget animations on data updates. Enabled by default.
+ * @param  {boolean} [props.global] - Enable/disable the viewport filtering in the data fetching.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  */
@@ -31,56 +30,26 @@ function FormulaWidget({
   joinOperation,
   formatter,
   animation,
+  global,
   onError,
   wrapperProps
 }) {
-  const isSourceReady = useSelector((state) =>
-    selectAreFeaturesReadyForSource(state, dataSource)
-  );
-  const { filters, filtersLogicalOperator } = useSourceFilters({ dataSource, id });
-
-  const [formulaData, setFormulaData] = useState(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    setIsLoading(true);
-
-    if (isSourceReady) {
-      getFormula({
-        operation,
-        column,
-        joinOperation,
-        filters,
-        filtersLogicalOperator,
-        dataSource
-      })
-        .then((data) => {
-          if (data && data[0]) {
-            setIsLoading(false);
-            setFormulaData(data[0].value);
-          }
-        })
-        .catch((error) => {
-          setIsLoading(false);
-          if (onError) onError(error);
-        });
-    }
-  }, [
-    operation,
-    column,
-    joinOperation,
-    filters,
-    filtersLogicalOperator,
+  const { data, isLoading } = useWidgetFetch(getFormula, {
+    id,
     dataSource,
-    setIsLoading,
-    onError,
-    isSourceReady
-  ]);
+    params: {
+      operation,
+      column,
+      joinOperation,
+    },
+    global,
+    onError
+  });
 
   return (
     <WrapperWidgetUI title={title} isLoading={isLoading} {...wrapperProps}>
       <FormulaWidgetUI
-        data={formulaData}
+        data={data}
         formatter={formatter}
         unitBefore={true}
         animation={animation}
@@ -99,12 +68,14 @@ FormulaWidget.propTypes = {
   operation: PropTypes.oneOf(Object.values(AggregationTypes)).isRequired,
   formatter: PropTypes.func,
   animation: PropTypes.bool,
+  global: PropTypes.bool,
   onError: PropTypes.func,
   wrapperProps: PropTypes.object
 };
 
 FormulaWidget.defaultProps = {
   animation: true,
+  global: false,
   wrapperProps: {}
 };