+++ packages/react-widgets/src/widgets/PieWidget.js
@@ -1,14 +1,13 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import React, { useCallback } from 'react';
+import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import { addFilter, removeFilter } from '@carto/react-redux';
 import { WrapperWidgetUI, PieWidgetUI, NoDataAlert } from '@carto/react-ui';
 import { _FilterTypes as FilterTypes, AggregationTypes } from '@carto/react-core';
 import { getCategories } from '../models';
-import useSourceFilters from '../hooks/useSourceFilters';
-import { selectAreFeaturesReadyForSource } from '@carto/react-redux/';
 import { useWidgetFilterValues } from '../hooks/useWidgetFilterValues';
 import { columnAggregationOn } from './utils/propTypesFns';
+import useWidgetFetch from '../hooks/useWidgetFetch';
 
 const EMPTY_ARRAY = [];
 
@@ -28,6 +27,7 @@ const EMPTY_ARRAY = [];
  * @param  {string} props.height - Height of the chart
  * @param  {boolean} [props.animation] - Enable/disable widget animations on data updates. Enabled by default.
  * @param  {boolean} [props.filterable] - Enable/disable widget filtering capabilities. Enabled by default.
+ * @param  {boolean} [props.global] - Enable/disable the viewport filtering in the data fetching.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  * @param  {Object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]()
@@ -46,6 +46,7 @@ function PieWidget({
   labels,
   animation,
   filterable,
+  global,
   colors,
   onError,
   wrapperProps,
@@ -53,54 +54,22 @@ function PieWidget({
 }) {
   const dispatch = useDispatch();
 
-  const [categoryData, setCategoryData] = useState([]);
   const selectedCategories =
     useWidgetFilterValues({ dataSource, id, column, type: FilterTypes.IN }) ||
     EMPTY_ARRAY;
-  const [isLoading, setIsLoading] = useState(true);
 
-  const isSourceReady = useSelector((state) =>
-    selectAreFeaturesReadyForSource(state, dataSource)
-  );
-  const { filters, filtersLogicalOperator } = useSourceFilters({ dataSource, id });
-
-  useEffect(() => {
-    setIsLoading(true);
-
-    if (isSourceReady) {
-      getCategories({
-        column,
-        operation,
-        operationColumn,
-        joinOperation,
-        filters,
-        filtersLogicalOperator,
-        dataSource
-      })
-        .then((data) => {
-          if (data) {
-            setIsLoading(false);
-            setCategoryData(data);
-          }
-        })
-        .catch((error) => {
-          setIsLoading(false);
-          if (onError) onError(error);
-        });
-    }
-  }, [
+  const { data = [], isLoading } = useWidgetFetch(getCategories, {
     id,
-    column,
-    operationColumn,
-    joinOperation,
-    operation,
-    filters,
-    filtersLogicalOperator,
     dataSource,
-    setIsLoading,
-    onError,
-    isSourceReady
-  ]);
+    params: {
+      column,
+      operationColumn,
+      joinOperation,
+      operation
+    },
+    global,
+    onError
+  });
 
   const handleSelectedCategoriesChange = useCallback(
     (categories) => {
@@ -128,9 +97,9 @@ function PieWidget({
 
   return (
     <WrapperWidgetUI title={title} isLoading={isLoading} {...wrapperProps}>
-      {categoryData.length || isLoading ? (
+      {data.length || isLoading ? (
         <PieWidgetUI
-          data={categoryData}
+          data={data}
           formatter={formatter}
           height={height}
           tooltipFormatter={tooltipFormatter}
@@ -165,6 +134,7 @@ PieWidget.propTypes = {
   labels: PropTypes.object,
   animation: PropTypes.bool,
   filterable: PropTypes.bool,
+  global: PropTypes.bool,
   onError: PropTypes.func,
   colors: PropTypes.arrayOf(PropTypes.string),
   wrapperProps: PropTypes.object,
@@ -174,6 +144,7 @@ PieWidget.propTypes = {
 PieWidget.defaultProps = {
   animation: true,
   filterable: true,
+  global: false,
   wrapperProps: {},
   noDataAlertProps: {}
 };